A lot of stuff
Created a new script called SudokuCreation which will be used only for the creation of the sudoku game on various challenge levels. It used Unity Job system to free the Main Thread from doing it.
We'll see if it'll work on the future.

I'll create a lot more scripts, reusing the previous codes, but making a cleaner code and scripts.